SM-1287: Finish refactor of project commands out of main.rs
