+++ crates/bws/src/command/project.rs
@@ -1,4 +1,12 @@
-use bitwarden::{secrets_manager::projects::ProjectsListRequest, Client};
+use std::process;
+
+use bitwarden::{
+    secrets_manager::projects::{
+        ProjectCreateRequest, ProjectGetRequest, ProjectPutRequest, ProjectsDeleteRequest,
+        ProjectsListRequest,
+    },
+    Client,
+};
 use bitwarden_cli::Color;
 use color_eyre::eyre::Result;
 use uuid::Uuid;
@@ -16,8 +24,100 @@ pub(crate) async fn list(
         .list(&ProjectsListRequest { organization_id })
         .await?
         .data;
-
     serialize_response(projects, output, color);
+
+    Ok(())
+}
+
+pub(crate) async fn get(
+    mut client: Client,
+    project_id: Uuid,
+    output: Output,
+    color: Color,
+) -> Result<()> {
+    let project = client
+        .projects()
+        .get(&ProjectGetRequest { id: project_id })
+        .await?;
+    serialize_response(project, output, color);
+
+    Ok(())
+}
+
+pub(crate) async fn create(
+    mut client: Client,
+    organization_id: Uuid,
+    name: String,
+    output: Output,
+    color: Color,
+) -> Result<()> {
+    let project = client
+        .projects()
+        .create(&ProjectCreateRequest {
+            organization_id,
+            name,
+        })
+        .await?;
+    serialize_response(project, output, color);
+
+    Ok(())
+}
+
+pub(crate) async fn edit(
+    mut client: Client,
+    organization_id: Uuid,
+    project_id: Uuid,
+    name: String,
+    output: Output,
+    color: Color,
+) -> Result<()> {
+    let project = client
+        .projects()
+        .update(&ProjectPutRequest {
+            id: project_id,
+            organization_id,
+            name,
+        })
+        .await?;
+    serialize_response(project, output, color);
+
+    Ok(())
+}
+
+pub(crate) async fn delete(mut client: Client, project_ids: Vec<Uuid>) -> Result<()> {
+    let count = project_ids.len();
+
+    let result = client
+        .projects()
+        .delete(ProjectsDeleteRequest { ids: project_ids })
+        .await?;
+
+    let projects_failed: Vec<(Uuid, String)> = result
+        .data
+        .into_iter()
+        .filter_map(|r| r.error.map(|e| (r.id, e)))
+        .collect();
+    let deleted_projects = count - projects_failed.len();
+
+    if deleted_projects > 1 {
+        println!("{} projects deleted successfully.", deleted_projects);
+    } else if deleted_projects == 1 {
+        println!("{} project deleted successfully.", deleted_projects);
+    }
+
+    if projects_failed.len() > 1 {
+        eprintln!("{} projects had errors:", projects_failed.len());
+    } else if projects_failed.len() == 1 {
+        eprintln!("{} project had an error:", projects_failed.len());
+    }
+
+    for project in &projects_failed {
+        eprintln!("{}: {}", project.0, project.1);
+    }
+
+    if !projects_failed.is_empty() {
+        process::exit(1);
+    }
 
     Ok(())
 }
